* Got rid of the fixed-width integers in the rnumber_t namespace 	and replaced them with the standard declarations in <stdint.h>. 	RNumber now compiles on 64-bit Linux, except for the Perl 	bindings, which have a link error.

+++ tests/NumberTest.C
@@ -897,7 +897,7 @@ static unsigned checkConstructors ()
   unsigned i;
   unsigned rc = 0;
 
-  unsigned long long x = 0x12345678deadbeefULL;
+  uint64_t x = 0x12345678deadbeefULL;
   RNumber a(x);
 //  cout << "x " << ::hex << x << " a " << a << endl;
